updated arduino file
You had the default random number coming in as value from the example arduino. this is currently my working arduino code that will buzz when the any heart rate is above 0. Change value of heartrate to what you think is the correct value of stressed heartrate.  Or change it to your working arduino code if you would like.